MAS-398 ensure scope is not set in oauth client config
