+++ src/main/resources/application.yml
@@ -25,13 +25,11 @@ spring:
             client-id: ${oauth.client.id}
             client-secret: ${oauth.client.secret}
             authorization-grant-type: client_credentials
-            scope: read
           hmpps-tier-api:
             provider: hmpps-auth
             client-id: ${oauth.client.id}
             client-secret: ${oauth.client.secret}
             authorization-grant-type: client_credentials
-            scope: read
         provider:
           hmpps-auth:
             token-uri: ${oauth.endpoint.url}/oauth/token
